requirements.txt
aiofiles==23.2.1
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.1.2
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
cryptography==42.0.5
ecdsa==0.18.0
exceptiongroup==1.2.0
fastapi>=0.110.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.5
httpx>=0.26.0
idna==3.6
iniconfig==2.0.0
packaging==24.0
passlib[bcrypt]>=1.7.4
pluggy==1.4.0
pyasn1==0.6.0
pycparser==2.22
pydantic>=2.6.1
pydantic-core==2.16.3
pydantic-settings>=2.1.0
pypng==0.20220715.0
pytest>=8.0.0
pytest-asyncio>=0.23.5
pytest-mock==3.14.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
qrcode>=7.4.2
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
tomli==2.0.1
typing_extensions==4.10.0
uvicorn>=0.27.1
validators>=0.24.0
